[llubi] Extract reusable methods of InstExecutor into ExecutorAPI
